+++ src/Components/Attendees.js
@@ -1,3 +1,7 @@
-export default function Attendees() {
-  return;
+export default function Attendees({toggleEventAttendees, showAttendees}) {
+  return (
+    <button onClick={toggleEventAttendees}>
+      {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+    </button>
+  );
 }
